refactor(phases): extract FileGroupBudget into its own module
- Move FileGroupBudget dataclass out of phases.py into new
  daydream/file_group_budget.py module (module-bloat ban; phases.py should
  not hold non-trivial dataclasses). Update imports in phases.py and
  tests/test_fix_budget.py.
- Extract the per-file-group budget resolution logic in _step_fix
  (deep/orchestrator.py) into three resolver helpers
  (_resolve_group_max_wall_s, _resolve_group_max_serial_items,
  _resolve_group_max_cumulative_tokens), mirroring the existing
  _shallow_fanout_threshold pattern. Behavior-preserving: same file-config
  precedence and is-not-None semantics.
- Remove the redundant pre-call budget.check() no-op in the batched path
  of phase_fix_parallel (a fresh group has spent no budget, so the check
  always passed; the comment documented it as a no-op).

Daydream-Run: 20260704160606-4ab71279
Daydream-Version: 0.20.0
